Isaacgarc-dev/issue47 (#62)
