+++ src/main/services/TSVService.ts
@@ -22,8 +22,9 @@ class TSVService {
       await writeFile(reportPath, content);
       return true;
     } catch (error) {
-      console.error("Error saving vendors to Excel file:", error);
-      return false;
+      const errorMessage = "Writting TSV to File\t";
+      // console.error(errorMessage, error);
+      throw errorMessage + "Failed Writing TSV to File. Not Directory Found.";
     }
   }
 }
